add show other users, increase size of icons
for now it has a temoorary button placed on top. when clicked, shows a
list of other users. when an user is clicked it will set a state that
will conditionally render either logged user or the selected user. when
this is selected, there is another temoorary button that lets you go
back to the logged user