handle multiple user visualization when logged out
previously it wouldn't reload the page with updated elements when logged
out or logged in iwhtout having to manually reload the page. that meant
that if i log in while im watching my own logged out profile, it woudl
still show as logged out until i reload the page. i sorted the issue by
allowing the page to sort between logged and non logged user when it
starts, but later in the page if an user logs in it should add that on
the list of user with a dirrerent click handler than a logged out user.
this sets the current user as such

now the page allows to show all the users, and when im logged in, i can
still navigate through users via the dropdown menu with all the users