add react-router, remove ShowUsers and ShowUsersList, create NavButton
instead of using conditional rendering with boolean states to decide
what should be on the screen, the navigational button has been moved in
his own file and everithing else is rendered inside a router. at the
start it will go to "/" which is currently emppty but it odesnt create
issues as navbar and navButton are outside of it so the user still has
options. when logged, route to /user, when clicked on the user list,
route to /other with that particular user. if we are logged in and the
uid matches the logged uid then go to /user.

2 files were removed as they were obsolete since this is regulated by
the router