attempt to fix data erasing bug
there is some bug that will erase the data of the user, but ive been
unable to replicate it as it is not clear where it gets effect. since
the use effects are fetching, then setting, then updating, its possible
that a delay in the fetch might set the empty object from the
initialized state. as an attempt, the useState have been put in order of
execution in the file. the bug is probably still unresolved