add toggle for en/it language
creates a context component with an object with all the reference of the
text used by the page written in italian and english. the toggle in the
navbar determine the state of a string which is then used in a switch to
pass the value to the LangContext.Provider. all components have been
updated to contain the value passed by the useContext.