resolve development bug that would erase my logged data
this was due to ItemListArrayComponent that would initialize the array
when found empty. this was caused in delays in the effect during
development that would see setData(with empty values) be faster than
fetchDatasetting resulting in the data being immediately set on
firestore when logged with just name, uid and photourl, then processed
by the component with empty arrays, then updated to firestore as soon as
user is modified.

to resolve this, setData and fetchData have been put inside the same
useEffect. if the user is logged, it will try to merge the docs with the
basic data, after that it will fetch the data from firestore. this puts
both process in the same hook and they will have to run one after the other instead of risking losing data due to delays.

it sets with merge option on true, then it fetches the data, then run any update if needed