remove the OtherUser folder and all his components, use ShowUser instead by typescript optional parameter, rename for remaining components
using the optional parameter in typescript when passing props, the
setUser hook has been put as optional and all the children have been
refactored to allow this change. this mean that now logged user and
other users can be both visualized with ShowUser instead of having
different components that resemple closely each other. ShowUser
component and all his children now allow for conditional rendering of
the options to modify the user given that props.setUser is truthy

added message in LangContextProvider to warn user that other profile
cannot be modified