add basic swipe action on Item.tsx
useSwipable is used to add a swipeable action on ech element. swiping left will show an absolutely positioned div in the center of the component, asking to confirm deletion (deletion TODO), while swiping  right will highlight the component with a different background color. this can be resetted by swiping again in the same direction, same for delete