create new toggle system for unavailable/available/ready
instead of having 2 buttons to declare if an item is unavailable (nothing is checked)
available (one is checked, one is not) and ready (all is checked) i decided to have
only one button on screen that will change is value everytime is clicked. a new value
is added to iGear, users with old values will be automaticallu updated by an use
effect that will determine if the vaule is "", in that case it will use a nested
ternary operator to put the latest value (if is available, check if is ready, if is
ready the put ready, otherwise available, otherwise unavailable). the icon is easy to read thanks to color changes (green/amber/red) and a text at the bottom