add full support for multiple travels
instead of passing the data as iUser, we pass it as iTravelData which contains userid
as well as travelid (this might need to be refactored later). needs more testing and
ironing of the UI. Stats component has been deactivated temporarely due to time
constraint while working, will be reinstated later.

the show user button is still showing 3 users instead of showing only the user for
that trip., this has to be investigated fully in the next session, for now it has been
left like this as it is not a breaking issue. its very likely due to the fact that the
users state adds indiscriminately from the list of users, since there is no rule to
discard users without the travel id yet.