fix offline and oflline data fetching bug
the if/else logic has been fixed to allow creation, modification and lists of data, by
refactoring the way this are made. more refactoring will be needed in the future. fix
the bug that wouldn't allow for the offline visualization to work (previoulsy it would
load the data if offline, it would work online) turned out that it was a mistake in
the way the travel.id was passed to the showUser component, as this was collected
traight when the user state was populated when logged in and then passed to the
component that would show the logged out lists. since user is not populated if logged out, the issue was simply that travelId was invalid when offline as it was undefined. this was solved by simply passing selectedTravel.id instead of user[selectedTravel.id].id. it's unclear how this happened