remove bug for createNewTravel, add Telegram messaging
with the new implementation came a bug that would prevent creating a new travel
object. the new travel in the past was using data from the user state to add
information to it, but since the new implementation wont create an user state until a
travel is clicked, this wouldnt work. the fix simply goeas aroudn it and pass loggedUser to get the data from it instead of user.

starting testing with telegram for another project. for now it will send to the bot
when a travel has been created or cancelled, will have more in the future (like, logged in users)