fix userTravelDataFactory function when creating newEmpty data set
there was an issue in the ternary operator that wouldnt create a new key in the object
with an empty array, instead it would look for the tempdata[delectedTravel.id].key.
this has been fixed by simplifying the ternary operator statement. in the past it was
lookign for two conditions (either collatedLoggedUser or collatedOterUser) to add the
tempdata key, now instead it sear3ch for the simplified options (if newEmpty) and if
true the value is [], otherwise tempdata