add handleClickOutsideComponent on USerButton.tsx
now the users button will close down automatically when the user click anywhere else on the screen. this has been done by creating a function that accepts as parameter the boolean state that determines the css display of the component. the ref has assigned the function that will determine the state