fix the view and join buttons
this has been archived by decoupling the old hook that would automatically fetch data when the user has logged in, as it would work perfectly for the old implementation. in the new system the hook has been removed and put in a function that will be passed down to the button and triggered on command. from now on, the only ative hooks are the one that fetch the travel list and update the data, while the fetching of the users is actually done on command when requested by the user, and not hooked to the log in