fix bug with DOM update and join/wiew function running late
the issue was that previous implementation were using useEffect hook heavily to update
the DOM after the data was retrieved, so the new implementation initally didn't take
account for the DOM delay when a state that dont render the current screen is updated.

what was happening, is that despite setting the selectedtravel correctly, the user
buttons would always show the previous state instead of the most current one (for
example, always the empty state at first, you had to click twice to geet the correct
data, then selecting a new data would still show the previous one), this is because
the dom wasnt updated in this case as the update would happen in a screen that hasn't
rendered yet. to fix this, the isWatching state has been passed down to view and Join button an is triggered together with handleclicksettravel. when this happens, an UseEffect in main that has selectedTravel as dependency will check on an if statement if isWatching is true or false, then fire the correct function (either watchtravel() or joinTravel())

the way the data was created to do so has also been changed: watchTravel doeesnt have
his own newTravelData object but it make use of usertraveldatafactory to create his
data, as its compatible since we have all the necessary data is in selectedTravel.
jointravel has been adapted for the latest changes, since we are not supposed to have
the button when the user is not logged, so there is no reason to have an else statement
to deal with that branch of joinTravel that used to deal with unlogged users that were
just watching the data. now an unlogged user only sees the watch button on screen so this change is reflected in the function