create deleteUserButton
the deleteUser function is stored in Main.tsx and passed down to User.tsc as argument
for the /user route. this means that it can be clicked only when user is logged in.
the function will get the user by UID, make a copy by value, delete the key with the
travelId, then set the user back on firestore. (as the hook tht updates the user only
updates for user[travelId] of type iUser not iTravelData so it wont fire
automatically). after doing so, it sets the selectedTravel to 0 (otherwise since join
and view are based on hook with selectedTravel as dependency, it wont fire further if
i want to keep doing some operation on that travel), then it will fire the
watchTravel(), as a quick way to refresh the usersList

the button has been styled minimally, will need more in the future