fix the swipe issue in Items.tsx
the issue was that swiping rx would interact witht he toggle button and create issues
with the main animation (button being stuck due to rendering). this has been fixed by
dividing the toggle logic and the return of the togglecomponent. the return virtualDom
items have been placed in the return of the Item component instead of being his own
component, the logic that changes and set the state of the component has been simply put on OnClick