test for Framer Motion library, fix Item.tsx isEditName conditional rendering
make several initial test with Framer Motion. for now it just makes a pop up animation
when the travelbutton are loaded, as well as giving a touch feedback when a single
travelbutton is clicked/touched. more to follow.

the name edit div has been modified: its not absolutely positioned anymore, but it
will replace the normal item div when requested. the cancel button has been removed to
make the UI clearer, a text has been added that say "confim" on the confirm button