refactor the webapp to have usersButtons to function properly
it creates an object called uiTriggers of type iTriggers that includes the state and
the setState "isShowUserButton" in order to decrease the amount of props drilling.
this system will be used to pass arund other boolean triggers. the new set state
simplify the sistem so that now the trigger is not a way to determine if we are
wathing or joining, but just a way to open the button list view.