refactr the login button and the createNewTravel.tsx
createNewTravel will now conditionally render to have the login button in its place,
this will disappear when logged and be replaved by the normal functionality to add a
new travel instead of being just missing from the screen when not logged. this way the
div shows a message for the user to log to create and join travels as well as the
login button. the navbar doesn't show the login propt anymore, but the logged user
microdashboard will appear as before in the same place. this has been done by setting
two optional props for the GoogleLoginButton.tsx, the toggle (which doesnt need to be
passed on the presentation screen, and the string that defines the message to show
when not logged). effectively there are now 2 GoogleLoginButtons on screen, but they are conditionally rendered off or on, so there is always just one on screen