refactor props drill, adds more to the lang functionality
travelButtonPropsPackage and userStatePropsPackage have been created as object
containing the props that need to be passed down the components.
teravelButtonPropsPackage handle the shared functionality between usersButton and the
travelItemButton (as clicking the travel button will open the usersButton, or join)
usersStatePropsPackage contains all the information to set user, users, loggedUsers et etc.

this simplify the passage of props down the line as i need only to set the type on the
interface,then only the object has to be passed instead of multiple items

langContextProvider has now enough functionality to handle almost 100% the change of
language onthe entire page