refactor joinTravel() and userDataFactoryFunction, adds comment on main.tsx
since the join travel button is rendered conditionally by checking if the user has a
key under that travelId, there is no reason to have the joinTravel Function doing
checks to figure out if the user is joining for the first time or if it's a logged out user watching the page or it's a logged user. it will always be an user joining for the first time, so  the function has been refactored.

userDatafactoryFunction has been slighlty simplified, since now (per above) we dont have 3 conditions to which we need to generate data, but just 2 (either joining or watching, as watching automatically decides on userButton if we are /user or /other)